fix: renamed components, added sidebar component directory, fixed video responsiveness
