+++ src/App.css
@@ -1,4 +1,3 @@
-
 .logo {
   height: 6em;
   padding: 1.5em;
@@ -36,24 +35,24 @@
 }
 .main-layout {
   display: grid;
-  grid-template-columns: 1fr 250px;
+  grid-template-columns: 1fr;
   height: 100vh;
   transition: grid-template-columns 0.5s ease-in-out;
+  overflow-x: hidden;
 }
 
 .main-content {
   padding: 20px;
   transition: width 0.5s ease-in-out;
-  display:grid;
+  display: grid;
   overflow-x: hidden;
-  align-items: stretch;
+  overflow-y: hidden;
+  align-items: center;
   justify-items: stretch;
-  gap:5px;
+  gap: 5px;
   object-fit: cover;
+  max-height: 100%;
 }
-
-
-
 
 .main-content.full-width {
   width: 100%;
@@ -68,7 +67,7 @@
 
 .sidebar.closed {
   visibility: hidden;
-  
+  display: none;
 }
 
 .toggle-sidebar {
@@ -78,15 +77,9 @@
   z-index: 1100;
 }
 
-.sidebar-video:hover{
-cursor: pointer;
-color:red;
-}
-
-
-
-.grid-container {
-
+.sidebar-video:hover {
+  cursor: pointer;
+  color: red;
 }
 
 .grid-item {
@@ -105,6 +98,3 @@ color:red;
   box-shadow: 0px 4px 8px rgba(0, 0, 0, 0.2); /* Shadow to lift the item visually */
   cursor: grabbing;
 }
-
-
-
