+++ src/components/VideoPlayerGrid.tsx
@@ -1,9 +1,8 @@
 import MainContent from "../layout/MainContent";
 import { useVideo } from "../contexts/VideoListContext";
 import VideoPlayerWrapper from "./VideoPlayerWrapper";
-// import { useEffect, useRef } from 'react';
 
-const VideoGrid = () => {
+const VideoPlayerGrid = () => {
   const { videos, volumes } = useVideo();
 
   const gridStyle = {
@@ -11,12 +10,12 @@ const VideoGrid = () => {
     gridTemplateRows: `repeat(${Math.ceil(
       videos.length / Math.ceil(Math.sqrt(videos.length)),
     )}, 1fr)`,
+    alignItems: "center",
   };
 
   return (
     <MainContent style={gridStyle}>
       {videos.map((video: string | null, index: number) => (
-        //need to add uniqu key to each video
         <VideoPlayerWrapper
           key={video}
           url={video || undefined}
@@ -28,6 +27,5 @@ const VideoGrid = () => {
   );
 };
 
-export default VideoGrid;
+export default VideoPlayerGrid;
 // api key = AIzaSyDww1kUVYoCUFQiSFQJZbGqdtkUbyCYnUs
-
